backend/requirements.txt
aiofiles==24.1.0
aiohttp==3.9.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
azure-ai-formrecognizer==3.1.0
azure-cognitiveservices-vision-computervision==0.7.0
azure-common==1.1.28
azure-core==1.32.0
azure-cosmos==4.5.0
azure-identity==1.15.0
azure-search-documents==11.4.0b6
azure-storage-blob==12.17.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
debugpy==1.8.1
distlib==0.3.8
distro==1.9.0
filelock==3.15.3
Flask==3.0.3
frozenlist==1.5.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
objprint==0.2.3
openai==1.55.3
orjson==3.10.1
packaging==24.1
platformdirs==4.2.2
portalocker==2.10.1
priority==2.0.0
propcache==0.2.1
psutil==6.0.0
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.2.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
python-dotenv==1.0.0
Quart==0.19.4
requests==2.32.3
rich==13.7.1
six==1.16.0
sniffio==1.3.1
subprocess32==3.5.4
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.24.0
virtualenv==20.26.2
vizplugins==0.1.3
viztracer==0.15.6
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3